Split up physics, removing all unused imports
