+++ src/illuminant/cie_data.rs
@@ -1,10 +1,8 @@
-use nalgebra::{ArrayStorage, SMatrix, SVector};
 
-use crate::{
-    illuminant::Illuminant,
-    observer::{Observer, ObserverData},
-    spectrum::{Spectrum, NS},
-};
+
+use nalgebra::{SVector, ArrayStorage};
+use crate::illuminant::Illuminant;
+use crate::spectrum::{Spectrum, NS};
 
 macro_rules! data_illuminant {
     ($name:ident, $data: tt) => {
