+++ src/illuminant/cie_illuminant.rs
@@ -1,10 +1,9 @@
 //! # CIE Standard Illuminants Enumerator
 
 use crate::{
-    error::Error, illuminant::Illuminant, spectrum::Spectrum, spectrum::NS, traits::Light,
+    illuminant::Illuminant, spectrum::Spectrum, traits::Light,
 };
-use nalgebra::{ArrayStorage, SMatrix};
-use std::{borrow::Cow, ops::Deref, vec};
+use std::borrow::Cow;
 use wasm_bindgen::prelude::*;
 
 macro_rules! std_illuminants {
