+++ src/illuminant/cri.rs
@@ -11,14 +11,7 @@ use std::{ops::Index, sync::LazyLock};
 use wasm_bindgen::prelude::*;
 
 use crate::{
-    colorant::Colorant,
-    error::Error,
-    illuminant::Illuminant,
-    observer::CIE1931,
-    rgb::RgbSpace,
-    spectrum::{Spectrum, SPECTRUM_WAVELENGTH_RANGE},
-    traits::Light,
-    xyz::XYZ,
+    colorant::Colorant, error::Error, illuminant::Illuminant, observer::CIE1931, spectrum::{Spectrum, SPECTRUM_WAVELENGTH_RANGE}, xyz::XYZ
 };
 
 /// Nummer of Test Color Sample Spectra
@@ -55,7 +48,7 @@ fn tcs_test() {
             &crate::illuminant::cie_illuminant::CieIlluminant::D65,
             Some(s),
         );
-        let [r, g, b] = xyz.rgb(Some(RgbSpace::SRGB)).values();
+        let [r, g, b] = xyz.rgb(Some(crate::rgb::RgbSpace::SRGB)).values();
         println!("{:2} ({r:3},{g:3},{b:3})", i + 1);
     }
