+++ src/lib.rs
@@ -1,10 +1,11 @@
-#![allow(dead_code, unused_variables, unused_imports)]
+#![allow(dead_code, unused_variables)]
 #![doc = include_str!("../README.md")]
 // This library defines many floating-point constants for colorimetry.
 // Clippy's `approx_constant` lint would otherwise generate numerous false positives
 // by flagging constants close to standard values.
 // To suppress these, `#![allow(clippy::approx_constant)]` is applied.
 #![allow(clippy::approx_constant)]
+///#![allow(dead_code, unused_variables, unused_imports)]
 
 pub mod cam;
 pub mod colorant;
