+++ src/observer.rs
@@ -47,12 +47,11 @@ mod observers;
 pub use observers::*;
 
 use crate::{
-    colorant::Colorant,
     error::Error,
     geometry::LineAB,
-    illuminant::CieIlluminant,
+    illuminant::{CieIlluminant, planck, planck_slope},
     lab::CieLab,
-    physics::{planck, planck_slope, to_wavelength},
+    physics::{to_wavelength},
     rgb::RgbSpace,
     spectrum::{Spectrum, NS, SPECTRUM_WAVELENGTH_RANGE},
     traits::{Filter, Light},
@@ -60,12 +59,11 @@ use crate::{
 };
 use nalgebra::{Matrix3, SMatrix, Vector3};
 use std::{
-    borrow::{Borrow, Cow},
-    ops::{Index, RangeInclusive},
+    ops::RangeInclusive,
     sync::OnceLock,
 };
 use strum_macros::EnumIter;
-use wasm_bindgen::{convert::IntoWasmAbi, prelude::wasm_bindgen};
+use wasm_bindgen::prelude::wasm_bindgen;
 
 /**
    Light-weight identifier added to the `XYZ` and `RGB` datasets,
@@ -160,7 +158,7 @@ impl ObserverData {
     ///
     /// The Light trait is implemented by [`CieIlluminant`] and [Illuminant](crate::illuminant::Illuminant).
     ///
-    /// [`Colorant`] implments the [`Filter`] trait.
+    /// [`Colorant`](crate::colorant::Colorant) implments the [`Filter`] trait.
     /// [`Rgb`](crate::rgb::Rgb), which represents a display pixel, implements both in this library.
     /// As a light, it is the light emitted from the pixel, as a filter it is the RGB-composite
     /// filter which is applied to the underlying standard illuminant of color space.
@@ -468,7 +466,7 @@ mod obs_test {
     use crate::prelude::{CieIlluminant, CIE1931};
     use crate::rgb::RgbSpace;
     use crate::spectrum::SPECTRUM_WAVELENGTH_RANGE;
-    use crate::xyz::{Chromaticity, XYZ};
+    use crate::xyz::XYZ;
     use approx::assert_ulps_eq;
     use strum::IntoEnumIterator as _;
 
@@ -605,7 +603,7 @@ mod obs_test {
 
     #[test]
     fn test_xyz_of_sample_with_standard_illuminant() {
-        use crate::prelude::{CieIlluminant::D65 as d65, XYZ};
+        use crate::prelude::{CieIlluminant::D65 as d65};
         let xyz = CIE1931
             .xyz(&d65, Some(&crate::colorant::Colorant::white()))
             .set_illuminance(100.0);
